Changes to the warning option
This warns the user so they sign in from only one device or browser at a time. If other sessions exist, the user receives a message and will have the option to terminate the other session and sign out. (The other sessions will not terminate automatically if you choose to sign in and continue to the new session)